fix: update reward definitions and fix unit test
