+++ sample/sample.ts
@@ -1,14 +1,23 @@
 
 import * as dotenv from 'dotenv'
 import { Omneo } from '../src'
+// import { Omneo, ID } from '../src'
+// import simpleOmneoRequest from '../src/test/lib/simple-omneo-request'
 
 dotenv.config()
 
-const { OMNEO_TOKEN: token, OMNEO_TENANT: tenant } = process.env
+const { OMNEO_TOKEN: token, OMNEO_TENANT: tenant, OMNEO_TEST_PROFILE_ID: testProfileId } = process.env
 
-const omneoClient = new Omneo({ token: token as string, tenant: tenant as string })
+const profileId = testProfileId as string
 
-console.log('Fetching Profiles')
-omneoClient.profiles.list().then((profiles) => {
-  console.log(`Retrieved ${profiles.length} profiles`)
-})
+function main () {
+  const omneoClient = new Omneo({ token: token as string, tenant: tenant as string })
+
+  console.log('profileId: ', profileId)
+  // const IDClient = new ID({ tenant: tenant as string, omneoAPIToken: token, config: {} })
+  // IDClient.auth.requestAuthToken({ id: testProfileId as string }).then((res) => {
+  //   console.log('response:', res)
+  // })
+}
+
+main()
